Finished api for qr code
